+++ support-frontend/assets/components/salesTax/maybeEstimatedTax.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 import type { CurrencyInfo } from '@modules/internationalisation/currency';
 import type { ReactNode } from 'react';
 import { simpleFormatTaxAmount } from 'helpers/forms/checkouts';
-import type { TaxRateConfig } from 'helpers/salesTax/getEstimatedSalesTaxRate';
+import type { TaxRateConfig } from 'helpers/salesTax/getEstimatedSalesTaxConfig';
 
 type Props = {
 	amount: number;
